move TPT Initiation class to cohorts package
